Small bugs and a helper method to be used in reports
